+++ common/lib/xmodule/xmodule/library_content_module.py
@@ -17,6 +17,7 @@ from webob import Response
 from xblock.core import XBlock
 from xblock.fields import Scope, String, List, Integer, Boolean
 from xblock.fragment import Fragment
+from xmodule.util.adaptive_learning import AdaptiveLearningAPIMixin
 from xmodule.validation import StudioValidationMessage, StudioValidation
 from xmodule.x_module import XModule, STUDENT_VIEW
 from xmodule.studio_editable import StudioEditableModule, StudioEditableDescriptor
@@ -390,99 +391,40 @@ class LibraryContentModule(LibraryContentFields, XModule, StudioEditableModule):
 
 @XBlock.wants('user')
 @XBlock.wants('library_tools')  # Only needed in studio
-class AdaptiveLibraryContentModule(AdaptiveLibraryContentFields, LibraryContentModule):
+class AdaptiveLibraryContentModule(AdaptiveLibraryContentFields, LibraryContentModule, AdaptiveLearningAPIMixin):
     """
     A specialized version of Randomized Content Block that communicates
     with an external service to determine when to show its children, and which ones.
     """
 
-    def __init__(self, descriptor, *args, **kwargs):
+    @lazy
+    def parent_course(self):
         """
-        Initialize cache for relevant data.
+        Return parent course.
         """
-        self._adaptive_learning_configuration = None
-        self._current_user_id = None
-        self._parent_unit_id = None
+        return self.runtime.modulestore.get_course(self.course_id, depth=1)
 
-        super(AdaptiveLibraryContentModule, self).__init__(descriptor, *args, **kwargs)
+    @lazy
+    def parent_unit_id(self):
+        """
+        Return block ID of parent unit.
+        """
+        parent_unit = self.get_parent()
+        return parent_unit.scope_ids.usage_id.block_id
 
-    @property
-    def adaptive_learning_configuration(self):
+    @lazy
+    def anonymous_user_id(self):
         """
-        Return configuration for accessing external service that provides adaptive learning features.
-
-        This configuration is a course-wide setting, so in order to access it,
-        we need to load the parent course from the DB.
-
-        We only do this once, on the first call.
-        On subsequent calls we return a cached value.
+        Return anonymous ID of current user.
         """
-        if self._adaptive_learning_configuration is None:
-            from courseware.courses import get_course_by_id
-            course = get_course_by_id(self.course_id, depth=1)
-            self._adaptive_learning_configuration = course.adaptive_learning_configuration
-        return self._adaptive_learning_configuration
-
-    @property
-    def adaptive_learning_url(self):
-        """
-        Return base URL for external service that provides adaptive learning features.
-
-        The base URL is a combination of the URL (url) and API version (api_version)
-        specified in the adaptive learning configuration for the parent course.
-        """
-        url = self.adaptive_learning_configuration.get('url')
-        api_version = self.adaptive_learning_configuration.get('api_version')
-        return '{url}/{api_version}'.format(url=url, api_version=api_version)
-
-    @property
-    def instance_url(self):
-        """
-        Return URL for requesting instance-specific data from external service
-        that provides adaptive learning features.
-        """
-        instance_id = self.adaptive_learning_configuration.get('instance_id')
-        return '{base_url}/instances/{instance_id}'.format(
-            base_url=self.adaptive_learning_url, instance_id=instance_id
-        )
-
-    @property
-    def students_url(self):
-        """
-        Return URL for requests dealing with students.
-        """
-        return '{base_url}/students'.format(base_url=self.instance_url)
-
-    @property
-    def event_url(self):
-        """
-        Return URL for requests dealing with events.
-        """
-        return '{base_url}/events'.format(base_url=self.instance_url)
-
-    @property
-    def knowledge_node_students_url(self):
-        """
-        Return URL for accessing 'knowledge node student' objects.
-        """
-        return '{base_url}/knowledge_node_students'.format(base_url=self.instance_url)
-
-    @property
-    def pending_reviews_url(self):
-        """
-        Return URL for accessing pending reviews.
-        """
-        return '{base_url}/review_utils/fetch_reviews'.format(base_url=self.instance_url)
-
-    @property
-    def request_headers(self):
-        """
-        Return custom headers for requests to external service that provides adaptive learning features.
-        """
-        access_token = self.adaptive_learning_configuration.get('access_token')
-        return {
-            'Authorization': 'Token token={access_token}'.format(access_token=access_token)
-        }
+        user_service = self.runtime.service(self, 'user')
+        if user_service:
+            current_user = user_service.get_current_user()
+            username = current_user.opt_attrs.get('edx-platform.username')
+            anonymous_user_id = user_service.get_anonymous_user_id(username, unicode(self.course_id))
+        else:
+            anonymous_user_id = None
+        return anonymous_user_id
 
     @classmethod
     def make_selection(cls, selected, children, max_count, mode):
@@ -540,33 +482,6 @@ class AdaptiveLibraryContentModule(AdaptiveLibraryContentFields, LibraryContentM
             'added': set(),
         }
 
-    def get_selections_current_user(self, children):
-        """
-        Check with external service whether any children of this block
-        are scheduled for review by the current user, and return them.
-        """
-        pending_reviews = self.get_pending_reviews_current_user()
-        valid_block_keys = {
-            c.block_id: (c.block_type, c.block_id) for c in children
-        }
-        selections = []
-        for pending_review in pending_reviews:
-            block_id = pending_review.get('knowledge_node_uid')
-            if block_id in valid_block_keys:
-                selections.append(valid_block_keys['block_id'])
-        return selections
-
-    def get_pending_reviews_current_user(self):
-        """
-        Return pending reviews for current user.
-        """
-        user_id = self.get_current_user_id()
-        url = self.pending_reviews_url
-        payload = {'student_uid': user_id}
-        response = requests.get(url, headers=self.request_headers, data=payload)
-        pending_reviews_user = json.loads(response.content)
-        return pending_reviews_user
-
     def selected_children(self):
         """
         Returns a set() of block_ids indicating which of the possible children
@@ -591,6 +506,22 @@ class AdaptiveLibraryContentModule(AdaptiveLibraryContentFields, LibraryContentM
 
         return selected
 
+    def get_selections_current_user(self, children):
+        """
+        Check with external service whether any children of this block
+        are scheduled for review by the current user, and return them.
+        """
+        pending_reviews = self.get_pending_reviews(self.anonymous_user_id)
+        valid_block_keys = {
+            c.block_id: (c.block_type, c.block_id) for c in children
+        }
+        selections = []
+        for pending_review in pending_reviews:
+            block_id = pending_review.get('knowledge_node_uid')
+            if block_id in valid_block_keys:
+                selections.append(valid_block_keys['block_id'])
+        return selections
+
     def student_view(self, context):
         """
         Render selected child blocks.
@@ -604,138 +535,12 @@ class AdaptiveLibraryContentModule(AdaptiveLibraryContentFields, LibraryContentM
         """
         Notify external service that parent unit has been viewed by learner.
         """
-        # Construct URL
-        url = self.event_url
-        # Construct event data
-        payload = {
-            'knowledge_node_student_id': self.get_knowledge_node_student_id(),
-            'event_type': 'EventRead',
-        }
-        # Send request
-        requests.post(url, headers=self.request_headers, data=payload)
-
-    def get_knowledge_node_student_id(self):
-        """
-        Return ID of 'knowledge node student' object linking current user and unit.
-        """
-        # Get student ID for current user
-        user_id = self.get_current_user_id()
         # Get block ID of parent unit
-        block_id = self.get_parent_unit_id()
-        # Get 'knowledge node student' object linking current user and parent unit
-        knowledge_node_student = self.get_or_create_knowledge_node_student(block_id, user_id)
-        return knowledge_node_student.get('id')
-
-    def get_current_user_id(self):
-        """
-        Return anonymous ID of current user.
-        """
-        if self._current_user_id is None:
-            user_service = self.runtime.service(self, 'user')
-            current_user = user_service.get_current_user()
-            username = current_user.opt_attrs.get('edx-platform.username')
-            self._current_user_id = user_service.get_anonymous_user_id(username, unicode(self.course_id))
-        return self._current_user_id
-
-    def get_parent_unit_id(self):
-        """
-        Return block ID of parent unit.
-        """
-        if self._parent_unit_id is None:
-            parent_unit = self.get_parent()
-            self._parent_unit_id = parent_unit.scope_ids.usage_id.block_id
-        return self._parent_unit_id
-
-    def get_or_create_knowledge_node_student(self, block_id, user_id):
-        """
-        Return 'knowledge node student' object for user identified by `user_id`
-        and unit identified by `block_id`.
-        """
-        # Create student
-        self.get_or_create_student(user_id)
-        # Link student to unit
-        knowledge_node_student = self.get_knowledge_node_student(block_id, user_id)
-        if knowledge_node_student is None:
-            knowledge_node_student = self.create_knowledge_node_student(block_id, user_id)
-        return knowledge_node_student
-
-    def get_or_create_student(self, user_id):
-        """
-        Create a new student on external service if it doesn't exist,
-        and return it.
-        """
-        student = self.get_student(user_id)
-        if student is None:
-            student = self.create_student(user_id)
-        return student
-
-    def get_student(self, user_id):
-        """
-        Return external information about student identified by `user_id`,
-        or None if external service does not know about student.
-        """
-        students = self.get_students()
-        try:
-            student = next(s for s in students if s.get('uid') == user_id)
-        except StopIteration:
-            student = None
-        return student
-
-    def get_students(self):
-        """
-        Return list of all students that external service knows about.
-        """
-        url = self.students_url
-        response = requests.get(url, headers=self.request_headers)
-        students = json.loads(response.content)
-        return students
-
-    def create_student(self, user_id):
-        """
-        Create student identified by `user_id` on external service,
-        and return it.
-        """
-        url = self.students_url
-        payload = {'uid': user_id}
-        response = requests.post(url, headers=self.request_headers, data=payload)
-        student = json.loads(response.content)
-        return student
-
-    def get_knowledge_node_student(self, block_id, user_id):
-        """
-        Return 'knowledge node student' object for user identified by `user_id`
-        and unit identified by `block_id`, or None if it does not exist.
-        """
-        # Get 'knowledge node student' objects
-        links = self.get_knowledge_node_students()
-        # Filter them by `block_id` and `user_id`
-        try:
-            link = next(
-                l for l in links if l.get('knowledge_node_uid') == block_id and l.get('student_uid') == user_id
-            )
-        except StopIteration:
-            link = None
-        return link
-
-    def get_knowledge_node_students(self):
-        """
-        Return list of all 'knowledge node student' objects for this course.
-        """
-        url = self.knowledge_node_students_url
-        response = requests.get(url, headers=self.request_headers)
-        links = json.loads(response.content)
-        return links
-
-    def create_knowledge_node_student(self, block_id, user_id):
-        """
-        Create 'knowledge node student' object that links student identified by `user_id`
-        to unit identified by `block_id`, and return it.
-        """
-        url = self.knowledge_node_students_url
-        payload = {'knowledge_node_uid': block_id, 'student_uid': user_id}
-        response = requests.post(url, headers=self.request_headers, data=payload)
-        knowledge_node_student = json.loads(response.content)
-        return knowledge_node_student
+        block_id = self.parent_unit_id
+        # Get student ID for current user
+        user_id = self.anonymous_user_id
+        # Send "Unit viewed" event
+        self.create_read_event(block_id, user_id)
 
 
 @XBlock.wants('user')
